+++ include/pfasst/pfasst.hpp
@@ -14,7 +14,7 @@ namespace pfasst {
   template<typename timeT>
   class PFASST : public Controller<timeT> {
 
-    void run(timeT dt, int nsteps) {
+    void run(timeT dt, size_t nsteps) {
 
     }
